Normalize the encoding macros, MPG123_ENC_8 now working just like MPG123_ENC_16. Normalize return values: int mpg123_read();
git-svn-id: svn://scm.orgis.org/mpg123/trunk@1031 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/libmpg123.c
@@ -496,7 +496,7 @@ int mpg123_decode_frame(mpg123_handle *mh, off_t *num, unsigned char **audio, si
 	return MPG123_ERR;
 }
 
-ssize_t mpg123_read(mpg123_handle *mh, unsigned char *out, size_t size, size_t *done)
+int mpg123_read(mpg123_handle *mh, unsigned char *out, size_t size, size_t *done)
 {
 	return mpg123_decode(mh, NULL, 0, out, size, done);
 }
